requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
scipy
python-docx